Support `git fz*` without supplying a subcmd
I often find I want to use `fzsha` in particular just to find a commit,
not to pass it to any subcommand afterwards, which usually looks like:

    $ git fzsha
    [error]
    $ git fzsha alsldfikjng
    [interactive fzf that I want]
    [git error no such subcommand 'keyboardmash']

so this commit just finally makes the first not an error, and does what
I want. It could probably just do the interactive thing and exit, but
sometimes having a record of the SHA might be helpful.