Make EasyMotion faster to use; more like vimium
Not that vimium necessarily works a better way by default, just that
EasyMotion is easy to configure to be ~consistent. (Anywhere in file
rather than just clickable places, but not really an equivalent to that
in a text file. First character to choose the line; second to choose
somewhere on it might be nice, but haven't found an option like that.)