Fix kb media control not always working
It happens because if I e.g. use the earphones' button to play/pause, it
bumps to the top of the list as MRU. (It's in the list in the first
place because I suppose it's a player by virtue of having a
microphone/being an input source?)

Earphones aren't really a player in this context, can't imagine that's
ever what I want, so disabling. It's unfortunate to tie it to the model
name, but I can't immediately see an alternative.