Fix display never ready (according to service)
`sway --get-socketpath` seems to be essentially looking for `$SWAYSOCK`,
which was never updated as we ran it in a loop in a single shell
session.

This commit fixes it by letting systemd handle the retries (better
anyway!) which therefore means it runs each time in a new process, which
allows the environment to have changed.